+++ src/cucu/environment.py
@@ -294,7 +294,7 @@ def after_step(ctx, step):
     # and this step has no substeps as in the reporting the substeps that
     # may actually do something on the browser take their own screenshots
     if ctx.browser is not None and ctx.current_step.has_substeps is False:
-        take_screenshot(ctx, step.name, label=f"After {step.name}")
+        take_screenshot(ctx, step, step.name, label=f"After {step.name}")
 
     # if the step has substeps from using `run_steps` then we already moved
     # the step index in the run_steps method and shouldn't do it here