+++ src/cucu/utils.py
@@ -23,7 +23,7 @@ from tenacity import retry as retrying
 from cucu import logger
 from cucu.browser.core import Browser
 from cucu.config import CONFIG
-from cucu.database.hooks import record_screenshot_artifact
+from cucu.database import hooks
 
 GHERKIN_TABLEFORMAT = TableFormat(
     lineabove=None,
@@ -215,13 +215,14 @@ def take_saw_element_screenshot(ctx, thing, name, index, element=None):
 
     take_screenshot(
         ctx,
+        ctx.current_step,
         ctx.current_step.name,
         label=f'{observed} {prefix}{thing} "{name}"',
         element=element,
     )
 
 
-def take_screenshot(ctx, step_name, label="", element=None):
+def take_screenshot(ctx, step, step_name, label="", element=None):
     screenshot_dir = os.path.join(
         ctx.scenario_dir, get_step_image_dir(ctx.step_index, step_name)
     )
@@ -271,6 +272,6 @@ def take_screenshot(ctx, step_name, label="", element=None):
     if CONFIG["CUCU_MONITOR_PNG"]:
         shutil.copyfile(filepath, CONFIG["CUCU_MONITOR_PNG"])
 
-    record_screenshot_artifact(ctx, filepath)
+    hooks.record_screenshot_artifact(ctx, step, filepath)
 
     CONFIG["__STEP_SCREENSHOT_COUNT"] += 1