+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project closely adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## 1.3.0
+- Change - add step param to take_screenshot
 - Add - database functionality to store and query test execution data
 - Add - support for hierarchical section steps in the database
 - Add - cucu db command for querying test results